Put changelog into file, post on PyPI
Having an easy to find changelog is important. We make it as visible as possible here by both having it in a separate file and posting it in the `long_description` on PyPI. It's now filled with the release notes that were on Github releases, which is a somewhat non-obvious place since most projects don't use them.